+++ apps/ngx-tinynodes/src/app/components/nav-bar/nav-bar.component.ts
@@ -1,8 +1,8 @@
 import { Component, Input } from '@angular/core';
-import { ApplicationService } from '../../store/app/application.service';
+import { AppService } from '../../store/app/application.service';
 
 @Component({
-  selector: 'ngx-nav-bar',
+  selector: 'tinynodes-nav-bar',
   templateUrl: 'nav-bar.component.html',
   styleUrls: ['nav-bar.component.scss']
 })
@@ -10,7 +10,7 @@ export class NavBarComponent {
   @Input()
   title = 'Navbar Component';
 
-  constructor(private readonly app: ApplicationService) {}
+  constructor(private readonly app: AppService) {}
 
   public toggleSidebar() {
     this.app.toggleSidebar();