+++ apps/ngx-tinynodes/src/app/store/app/application.model.ts
@@ -1,3 +1,11 @@
+import { RouterLink } from '@angular/router';
+
+export interface MenuItem {
+  label: string;
+  link: RouterLink;
+}
+
 export interface ApplicationState {
   hidden: boolean;
+  menu: MenuItem[];
 }