+++ libs/ngx-editorjs-demo/src/lib/config/index.ts
@@ -1,11 +1,22 @@
-import { PluginCode } from '../plugins/code/code.plugin';
 import { PluginConfig } from '@tinynodes/ngx-editorjs';
-import { PluginSimpleImage } from '../plugins/simple-image/simple-image.plugin';
-import { PluginLink } from '../plugins/link/link.plugin';
-import { PluginMarker } from '../plugins/marker/marker.plugin';
+import {
+  PluginCode,
+  PluginHeader,
+  PluginLink,
+  PluginList,
+  PluginMarker,
+  PluginParagraph,
+  PluginSimpleImage
+} from '@tinynodes/ngx-editorjs-plugins';
 
-export function createTools(): PluginConfig {
+/**
+ * Creates the tools for the `ngx-tinynodes` demo application
+ */
+export function createNgxEditorJSDemoTools(): PluginConfig {
   return {
+    parargraph: new PluginParagraph(),
+    header: new PluginHeader(),
+    list: new PluginList(),
     code: new PluginCode(),
     image: new PluginSimpleImage(),
     link: new PluginLink(),