+++ libs/ngx-editorjs-demo/src/lib/ngx-editorjs-demo.module.ts
@@ -1,23 +1,24 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
+import { MatButtonModule, MatCardModule } from '@angular/material';
 import { RouterModule } from '@angular/router';
-import EditorJSCode from '@editorjs/code';
-import EditorJSImage from '@editorjs/image';
-import EditorJSList from '@editorjs/list';
-import EditorJSMarker from '@editorjs/marker';
 import { NgxEditorJSModule, UserPlugins } from '@tinynodes/ngx-editorjs';
+import {
+  PluginCodeModule,
+  PluginHeaderModule,
+  PluginLinkModule,
+  PluginListModule,
+  PluginMarkerModule,
+  PluginParagraphModule,
+  PluginSimpleImageModule
+} from '@tinynodes/ngx-editorjs-plugins';
 import { MarkdownModule } from 'ngx-markdown';
+import { createNgxEditorJSDemoTools } from './config';
 import { EditorPageComponent } from './containers/editor-page/editor-page.component';
 import { PageContainerComponent } from './containers/page-container/page-container.component';
 import { ngxEditorjsDemoRoutes } from './ngx-editorjs-demo.routes';
 import { PageStoreModule } from './store/pages/pages.module';
-import { CodeModule } from './plugins/code/code.module';
-import { createTools } from './config';
-import { SimpleImageModule } from './plugins/simple-image/simple-image.module';
-import { LinkModule } from './plugins/link/link.module';
-import { MarkerModule } from './plugins/marker/marker.module';
-import { MatCardModule, MatButtonModule } from '@angular/material';
 
 @NgModule({
   imports: [
@@ -27,17 +28,20 @@ import { MatCardModule, MatButtonModule } from '@angular/material';
     ReactiveFormsModule,
     PageStoreModule,
     NgxEditorJSModule,
-    CodeModule,
-    SimpleImageModule,
-    LinkModule,
-    MarkerModule,
+    PluginHeaderModule,
+    PluginParagraphModule,
+    PluginListModule,
+    PluginCodeModule,
+    PluginSimpleImageModule,
+    PluginLinkModule,
+    PluginMarkerModule,
     MatCardModule,
     MatButtonModule
   ],
   providers: [
     {
       provide: UserPlugins,
-      useFactory: createTools
+      useFactory: createNgxEditorJSDemoTools
     }
   ],
   declarations: [EditorPageComponent, PageContainerComponent],