+++ libs/ngx-editorjs-demo/src/lib/plugins/code/code.module.ts
@@ -1,7 +1,0 @@
-import { NgModule } from '@angular/core';
-import { PluginCode } from './code.plugin';
-
-@NgModule({
-  providers: [PluginCode]
-})
-export class CodeModule {}