+++ libs/ngx-editorjs-demo/src/lib/plugins/simple-image/simple-image.module.ts
@@ -1,7 +1,0 @@
-import { NgModule } from '@angular/core';
-import { PluginSimpleImage } from './simple-image.plugin';
-
-@NgModule({
-  providers: [PluginSimpleImage]
-})
-export class SimpleImageModule {}