+++ libs/ngx-editorjs/src/index.ts
@@ -1,14 +1,7 @@
 export { NgxEditorJSComponent } from './lib/containers/editorjs-component/editorjs.component';
 export { NgxEditorJSDirective } from './lib/directives/ngx-editorjs.directive';
 export { NgxEditorJSModule } from './lib/ngx-editorjs.module';
-export { PluginHeaderModule } from './lib/plugins/header/header.module';
-export { PluginHeader } from './lib/plugins/header/header.plugin';
-export { PluginListModule } from './lib/plugins/list/list.module';
-export { PluginList } from './lib/plugins/list/list.plugin';
-export { PluginParagraphModule } from './lib/plugins/paragraph/paragraph.module';
-export { PluginParagraph } from './lib/plugins/paragraph/paragraph.plugin';
 export { NgxEditorJSService } from './lib/services/editorjs.service';
 export { Block, HeaderBlock, ImageBlock, ListBlock, ListStyle, ParagraphBlock } from './lib/types/blocks';
 export { EditorJSConfig, NgxEditorJSConfig, NGX_EDITORJS_CONFIG } from './lib/types/config';
-export { BasePlugin, PluginConfig, UserPlugins } from './lib/types/plugins';
-export { PluginService as ɵPluginService } from './lib/services/plugins.service';
+export { BasePlugin, UserPlugins, PluginConfig } from './lib/types/plugins';