+++ libs/ngx-editorjs/src/lib/ngx-editorjs.module.ts
@@ -1,29 +1,63 @@
 import { CommonModule } from '@angular/common';
-import { ModuleWithProviders, NgModule, Optional, SkipSelf } from '@angular/core';
+import { InjectionToken, ModuleWithProviders, NgModule, Optional, SkipSelf } from '@angular/core';
 import { NgxEditorJSComponent } from './containers/editorjs-component/editorjs.component';
 import { NgxEditorJSDirective } from './directives/ngx-editorjs.directive';
-import { PluginHeaderModule } from './plugins/header/header.module';
-import { PluginListModule } from './plugins/list/list.module';
 import { NgxEditorJSService } from './services/editorjs.service';
-import { PluginService } from './services/plugins.service';
-import { NgxEditorJSConfig, NGX_EDITORJS_CONFIG } from './types/config';
-import { InitialPlugins, PluginConfig } from './types/plugins';
-import { PluginHeader } from './plugins/header/header.plugin';
-import { PluginList } from './plugins/list/list.plugin';
-import { PluginParagraphModule } from './plugins/paragraph/paragraph.module';
-import { PluginParagraph } from './plugins/paragraph/paragraph.plugin';
+import { NgxEditorJSPluginService } from './services/plugins.service';
+import { EditorJSConfig, NgxEditorJSConfig, NGX_EDITORJS_CONFIG } from './types/config';
 
-export function defaultConfig() {
+/**
+ * The default holder ID to attach `EditorJS` to
+ */
+const DEFAULT_HOLDER_ID = 'editor-js';
+
+/**
+ * Creates a configuration for EditorJS
+ * @param config Optional module configurations
+ */
+export function createConfig(config?: NgxEditorJSConfig): NgxEditorJSConfig {
+  if (!config || !config.editorjs) {
+    return {
+      editorjs: {
+        holder: DEFAULT_HOLDER_ID
+      }
+    };
+  }
+  const editorJsConfig: EditorJSConfig = {};
+  if (config.editorjs.autofocus) {
+    editorJsConfig.autofocus = config.editorjs.autofocus;
+  }
+  if (config.editorjs.data) {
+    editorJsConfig.data = config.editorjs.data;
+  }
+  if (config.editorjs.hideToolbar) {
+    editorJsConfig.hideToolbar = config.editorjs.hideToolbar;
+  }
+  if (config.editorjs.initialBlock) {
+    editorJsConfig.initialBlock = config.editorjs.initialBlock;
+  }
+  if (config.editorjs.minHeight) {
+    editorJsConfig.minHeight = config.editorjs.minHeight;
+  }
+  if (config.editorjs.placeholder) {
+    editorJsConfig.placeholder = config.editorjs.placeholder;
+  }
+  if (config.editorjs.sanitizer) {
+    editorJsConfig.sanitizer = config.editorjs.sanitizer;
+  }
+
   return {
-    editorjs: { autofocus: false, initialBlock: 'paragraph', holder: 'editor-js' }
+    editorjs: {
+      holder: config.editorjs.holder || DEFAULT_HOLDER_ID,
+      ...editorJsConfig
+    }
   };
 }
 
-export function createTools(): PluginConfig {
-  return { header: new PluginHeader(), list: new PluginList(), paragraph: new PluginParagraph() };
-}
-
-const DEFAULT_CONFIG: NgxEditorJSConfig = defaultConfig();
+/**
+ * Internal token for injecting the `NgxEditorJSConfig` into the config factory
+ */
+export const FOR_ROOT_OPTIONS_TOKEN = new InjectionToken<NgxEditorJSConfig>('forRoot() NgxEditorJSConfig.');
 
 /**
  * The `@tinynodes/ngx-editorjs` module provides a collection of features to allow
@@ -34,24 +68,10 @@ const DEFAULT_CONFIG: NgxEditorJSConfig = defaultConfig();
  * default instance.
  */
 @NgModule({
-  imports: [CommonModule, PluginHeaderModule, PluginListModule, PluginParagraphModule],
+  imports: [CommonModule],
   declarations: [NgxEditorJSComponent, NgxEditorJSDirective],
-  exports: [
-    NgxEditorJSComponent,
-    NgxEditorJSDirective,
-    PluginHeaderModule,
-    PluginParagraphModule,
-    PluginParagraphModule
-  ],
-  providers: [
-    NgxEditorJSService,
-    {
-      provide: InitialPlugins,
-      useFactory: createTools
-    },
-    PluginService,
-    PluginHeaderModule
-  ]
+  exports: [NgxEditorJSComponent, NgxEditorJSDirective],
+  providers: [NgxEditorJSService, NgxEditorJSPluginService]
 })
 export class NgxEditorJSModule {
   constructor(
@@ -65,20 +85,18 @@ export class NgxEditorJSModule {
    * which sets some defaults, or use the provided defaults.
    * @param config The optional configuration to pass
    */
-  static forRoot(config: NgxEditorJSConfig = DEFAULT_CONFIG): ModuleWithProviders {
+  static forRoot(@Optional() config?: NgxEditorJSConfig): ModuleWithProviders {
     return {
       ngModule: NgxEditorJSModule,
       providers: [
         {
+          provide: FOR_ROOT_OPTIONS_TOKEN,
+          useValue: config
+        },
+        {
           provide: NGX_EDITORJS_CONFIG,
-          useValue: {
-            editorjs: {
-              autofocus: config.editorjs.autofocus || DEFAULT_CONFIG.editorjs.autofocus,
-              holder: config.editorjs.holder || DEFAULT_CONFIG.editorjs.holder,
-              initialBlock: config.editorjs.initialBlock || DEFAULT_CONFIG.editorjs.initialBlock,
-              data: config.editorjs.data
-            }
-          }
+          useFactory: createConfig,
+          deps: [FOR_ROOT_OPTIONS_TOKEN]
         }
       ]
     };