+++ libs/ngx-editorjs/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2019-05-13
+
+### Added
+
+- Added all configuration options for `EditorJS` to the module configuration
+
+### Changed
+
+- Removed all plugins from the `ngx-editorjs` project. All plugins are now provided by the `@tinynodes/ngx-editorjs-plugins` module and a default set of tools must be provided in an application.
+- Added `NgxEditorJSPluginService` back to core, all plugins must now be provided with a `UserPlugin` provider.
+
 ## [1.1.2] - 2019-05-11
 
 ### Added